start semester
git-svn-id: https://ejava.info/repos/class/trunk/ejava-javaee@272 6a96efd0-f11b-0410-be78-92d216eb0a58

+++ jdbcDAO/src/main/resources/ddl/daoBook_drop.ddl
@@ -1,3 +1,3 @@
-drop table DAO_BOOK;
-drop table DAO_BOOK_UID;
-drop sequence DAO_BOOK_SEQ;
+drop table DAO_BOOK if EXISTS;
+drop table DAO_BOOK_UID if EXISTS;
+drop sequence DAO_BOOK_SEQ if EXISTS;
